refactor: shuffle exports
